backend/requirements.txt
asgiref==3.2.7
Django==3.0.5
django-filter==2.2.0
djangorestframework==3.11.0
Pillow==7.1.2
pytz==2019.3
sqlparse==0.3.1
